018 mutlithreadig work + instaloader_iterator

+++ src/python/telegram_bot.py
@@ -1,10 +1,11 @@
 from typing import List
+from instaloader import Instaloader
 from telebot import telebot, types
 from telebot.types import Message, InlineKeyboardMarkup
 import configparser
 from instaloader_api import Loader
-from utils import valid_username, files_handler, create_text_menu, get_start_text
-from dtos import ProfileResponse
+from instaloader_iterator import InstaloaderIterator
+from utils import valid_username, create_text_menu, get_start_text
 import time
 from database_service import Service
 from concurrent.futures import ThreadPoolExecutor
@@ -12,13 +13,25 @@ from concurrent.futures import ThreadPoolExecutor
 properties = configparser.ConfigParser()
 properties.read('/home/evgeniy/PycharmProjects/insta-bot/src/resources/application.properties')
 
+def loaders_init() -> InstaloaderIterator:
+    user_1 = properties['INSTAGRAM']['USER_1']
+    user_2 = properties['INSTAGRAM']['USER_2']
+    session_token_1 = properties['PATHS']['PATH_OS'] + 'src/resources/session-token-1'
+    session_token_2 = properties['PATHS']['PATH_OS'] + 'src/resources/session-token-2'
+    instaloader_1 = Instaloader()
+    instaloader_1.load_session_from_file(user_1, session_token_1)
+    instaloader_2 = Instaloader()
+    instaloader_2.load_session_from_file(user_2, session_token_2)
+    return InstaloaderIterator([instaloader_1, instaloader_2])
+
+INSTALOADERS = loaders_init()
 BOT = telebot.TeleBot(properties['TELEGRAM']['BOT'])
 SERVICE = Service(properties)
 EXECUTOR = ThreadPoolExecutor()
-LOADERS = {int(properties['TELEGRAM']['ADMIN_ID']): Loader(properties, BOT, SERVICE, EXECUTOR),
-           int(properties['TELEGRAM']['ANNA_ID']): Loader(properties, BOT, SERVICE, EXECUTOR),
-           int(properties['TELEGRAM']['DASHA_ID']): Loader(properties, BOT, SERVICE, EXECUTOR)}
-BOT.send_message(properties['TELEGRAM']['ADMIN_ID'], text='✅ INSTALOADER start')
+LOADERS = {int(properties['TELEGRAM']['ADMIN_ID']): Loader(properties, INSTALOADERS, BOT, SERVICE, EXECUTOR),
+           int(properties['TELEGRAM']['ANNA_ID']): Loader(properties, INSTALOADERS, BOT, SERVICE, EXECUTOR),
+           int(properties['TELEGRAM']['DASHA_ID']): Loader(properties, INSTALOADERS, BOT, SERVICE, EXECUTOR)}
+BOT.send_message(properties['TELEGRAM']['ADMIN_ID'], text='✅ INSTABOT start')
 
 
 @BOT.message_handler(commands=['start'])
@@ -80,21 +93,25 @@ def get_menu_markup(message: Message, mode: str, usernames: List[str] | None = N
 def change_mode(callback_query):
     loader = LOADERS.get(callback_query.message.chat.id)
     if loader:
-        mode = callback_query.data.split('|')[1]
+        if not loader.LOCK.is_set():
+            loader.LOCK.set()
+            mode = callback_query.data.split('|')[1]
 
-        if mode == 'query': mode = 'analyzeNew'
-        elif mode == 'analyzeNew': mode = 'remove'
-        elif mode == 'remove': mode = 'query'
+            if mode == 'query': mode = 'analyzeNew'
+            elif mode == 'analyzeNew': mode = 'remove'
+            elif mode == 'remove': mode = 'query'
 
-        markup = get_menu_markup(callback_query.message, mode)
+            markup = get_menu_markup(callback_query.message, mode)
 
-        if markup:
-            text = create_text_menu(mode)
-            BOT.edit_message_text(text, callback_query.message.chat.id, callback_query.message.message_id,
-                                  parse_mode='HTML', reply_markup=markup)
-        else:
-            text = '🧐 Здесь будет меню, когда ты сделаешь хотя бы 1 запрос'
-            BOT.edit_message_text(text, callback_query.message.chat.id, callback_query.message.message_id)
+            if markup:
+                text = create_text_menu(mode)
+                BOT.edit_message_text(text, callback_query.message.chat.id, callback_query.message.message_id,
+                                      parse_mode='HTML', reply_markup=markup)
+            else:
+                text = '🧐 Здесь будет меню, когда ты сделаешь хотя бы 1 запрос'
+                BOT.edit_message_text(text, callback_query.message.chat.id, callback_query.message.message_id)
+
+            loader.LOCK.clear()
     else:
         BOT.send_message(callback_query.message.chat.id, text='Нет доступа')
         BOT.send_message(properties['TELEGRAM']['ADMIN_ID'], text='Левый пользователь. Лог:\n\n' + str(callback_query.message))
@@ -109,10 +126,7 @@ def read_message(message):
             if not loader.LOCK.is_set():
                 loader.LOCK.set()
                 status_bar = BOT.send_message(message.chat.id, text='Делаю запрос...')
-                future = EXECUTOR.submit(lambda: loader.search_profile(username, status_bar))
-                response = future.result()
-                query_handler(response, message)
-                loader.LOCK.clear()
+                EXECUTOR.submit(lambda: loader.search_profile(username, status_bar))
         else:
             BOT.send_message(message.chat.id, text=f'❌ "{message.text}" - некорректный никнейм')
     else:
@@ -128,96 +142,22 @@ def query(callback_query):
             loader.LOCK.set()
             status_bar = BOT.send_message(callback_query.message.chat.id, text='Делаю запрос...')
             username = callback_query.data.split('|')[1]
-            response = loader.search_profile(username, status_bar)
-            query_handler(response, callback_query.message)
-            loader.LOCK.clear()
+            EXECUTOR.submit(lambda: loader.search_profile(username, status_bar))
     else:
         BOT.send_message(callback_query.message.chat.id, text='Нет доступа')
         BOT.send_message(properties['TELEGRAM']['ADMIN_ID'], text='Левый пользователь. Лог:\n\n' + str(callback_query.message))
 
 
-def query_handler(response: ProfileResponse, message: Message):
-    if response.type == 'private':
-        with open(response.avatar_path, 'rb') as photo:
-            BOT.send_photo(message.chat.id, photo, caption=response.text_message, parse_mode='HTML')
-    if response.type == 'no_stories':
-        markup = types.InlineKeyboardMarkup()
-        text_data = f'query|{response.username}'
-        markup.add(types.InlineKeyboardButton(text='🔍 Сделать новый запрос', callback_data=text_data))
-        with open(response.avatar_path, 'rb') as photo:
-            BOT.send_photo(message.chat.id, photo, caption=response.text_message,
-                           parse_mode='HTML', reply_markup=markup)
-    if response.type == 'has_stories':
-        markup = types.InlineKeyboardMarkup()
-        text_data = f'analyze|{response.username}|{int(time.time())}'
-        markup.add(types.InlineKeyboardButton(text='🐝 Прошерстить', callback_data=text_data))
-        with open(response.avatar_path, 'rb') as photo:
-            BOT.send_photo(message.chat.id, photo, caption=response.text_message,
-                           parse_mode='HTML', reply_markup=markup)
-    if response.type == 'error':
-        BOT.send_message(message.chat.id, text=response.text_message)
-
-
 @BOT.callback_query_handler(func=lambda call: call.data.startswith('analyze'))
-def analyze(callback_query):
+def hornet(callback_query):
     loader = LOADERS.get(callback_query.message.chat.id)
     if loader:
         if not loader.LOCK.is_set():
             loader.LOCK.set()
             callback_type, username, time_create = callback_query.data.split('|')
             status_bar = BOT.send_message(callback_query.message.chat.id, text='Загружаю сторис...')
-            response = loader.download_stories(username, status_bar, time_create)
-            text_message: str
-            if response.type == 'has_stories':
-                for story_data in response.story_data_array:
-                    if story_data.content == 'photo':
-                        with open(story_data.path, 'rb') as photo:
-                            BOT.send_photo(callback_query.message.chat.id, photo)
-                    if story_data.content == 'video':
-                        with open(story_data.path, 'rb') as video:
-                            BOT.send_video(callback_query.message.chat.id, video)
-
-                files_handler(response.story_data_array, response.folder_stories)
-
-                if not response.story_data_array:
-                    if response.count_stories == 1:
-                        text_message = (f'<code>Инсташершень:</code>\n\n'
-                                        f'У <b>{response.full_name}</b> сейчас одна актуальная сторис.\n'
-                                        f'Я тебе уже отправлял эту сторис - попробуй прошерстить этот аккаунт позже')
-                    else:
-                        text_message = (f'<code>Инсташершень:</code>\n\n'
-                                        f'У <b>{response.full_name}</b> {response.count_stories} актуальных сторис.\n'
-                                        f'Я тебе уже отправлял все эти сторис - попробуй прошерстить этот аккаунт позже')
-                elif response.count_viewed > 0:
-                    text_message = (f'<code>Инсташершень:</code>\n\n'
-                                    f'У <b>{response.full_name}</b> {response.count_stories} актуальных сторис.\n'
-                                    f'Я тебе отправил всего {len(response.story_data_array)} сторис, '
-                                    f'т.к. другие я тебе отправлял ранее')
-                else:
-                    text_message = (f'<code>Инсташершень:</code>\n\n'
-                                    f'Все актуальные сторис <b>{response.full_name}</b> отправлены')
-                BOT.send_message(callback_query.message.chat.id, text=text_message, parse_mode='HTML')
-            if response.type == 'no_stories':
-                text_message = (f'<code>Инсташершень:</code>\n\n'
-                                f'У <b>{response.full_name}</b> сейчас нет актуальных сторис, попробуй прошерстить его позже')
-                BOT.send_message(callback_query.message.chat.id, text=text_message, parse_mode='HTML')
-            if response.type == 'error_loader':
-                text_message = '❌ В данный момент нет ответа от instagram, попробуй сделать запрос позже — через 15-20 минут.'
-                BOT.send_message(callback_query.message.chat.id, text_message)
-
-
-            if callback_type == 'analyze':
-                markup = types.InlineKeyboardMarkup(row_width=2)
-                btn1 = types.InlineKeyboardButton(text='🔍 Новый запрос', callback_data=f'query|{username}')
-                btn2 = types.InlineKeyboardButton(text='🐝 Прошерстить',
-                                                  callback_data=f'analyzeNew|{username}|{int(time.time())}')
-                markup.add(btn1, btn2)
-                BOT.edit_message_reply_markup(
-                    chat_id=callback_query.message.chat.id,
-                    message_id=callback_query.message.message_id,
-                    reply_markup=markup
-                )
-            loader.LOCK.clear()
+            EXECUTOR.submit(lambda: loader.download_stories(callback_type, username, status_bar,
+                                                            callback_query.message, time_create))
     else:
         BOT.send_message(callback_query.message.chat.id, text='Нет доступа')
         BOT.send_message(properties['TELEGRAM']['ADMIN_ID'], text='Левый пользователь. Лог:\n\n' + str(callback_query.message))
@@ -227,17 +167,21 @@ def analyze(callback_query):
 def remove_history(callback_query):
     loader = LOADERS.get(callback_query.message.chat.id)
     if loader:
-        telegram_id = callback_query.message.chat.id
-        mode, username, time_created = callback_query.data.split('|')
-        refresh_usernames = SERVICE.remove_profile(telegram_id, username)
-        markup = get_menu_markup(callback_query.message, mode, refresh_usernames)
+        if not loader.LOCK.is_set():
+            loader.LOCK.set()
+            telegram_id = callback_query.message.chat.id
+            mode, username, time_created = callback_query.data.split('|')
+            refresh_usernames = SERVICE.remove_profile(telegram_id, username)
+            markup = get_menu_markup(callback_query.message, mode, refresh_usernames)
 
-        if markup:
-            text = create_text_menu(mode)
-            BOT.edit_message_text(text, callback_query.message.chat.id, callback_query.message.message_id, reply_markup=markup)
-        else:
-            text = '🧐 Здесь будет меню, когда ты сделаешь хотя бы 1 запрос'
-            BOT.edit_message_text(text, callback_query.message.chat.id, callback_query.message.message_id)
+            if markup:
+                text = create_text_menu(mode)
+                BOT.edit_message_text(text, callback_query.message.chat.id, callback_query.message.message_id, reply_markup=markup)
+            else:
+                text = '🧐 Здесь будет меню, когда ты сделаешь хотя бы 1 запрос'
+                BOT.edit_message_text(text, callback_query.message.chat.id, callback_query.message.message_id)
+
+            loader.LOCK.clear()
     else:
         BOT.send_message(callback_query.message.chat.id, text='Нет доступа')
         BOT.send_message(properties['TELEGRAM']['ADMIN_ID'], text='Левый пользователь. Лог:\n\n' + str(callback_query.message))
